Move christmas event sidebar out of folder
