+++ src/modules/events/views/ui/layouts/types/onsite/christmas-onsite-2020/ChristmasOnsite2020Sidebar.tsx
@@ -6,7 +6,7 @@ import Link from "next/link";
 import { Paths } from "@/app/Paths";
 import Spinner from "@/core/views/ui/app/Spinner";
 
-import { EventLayoutProps } from "../../../../layouts";
+import { EventLayoutProps } from "../../../layouts";
 import ChristmasOnsite2020SidebarPlaceholder from "./ChristmasOnsite2020SidebarPlaceholder";
 import ChristmasOnsite2020SidebarRoomList from "./ChristmasOnsite2020SidebarRoomList";
 import { formatDescription } from "./ChristmasOnsite2020SidebarUtils";
