+++ src/modules/events/views/ui/layouts/types/onsite/christmas-onsite-2020/ChristmasOnsite2020SidebarRoomList.tsx
